Move spectrum_pair_selection.py to train_ms2deepscore dir
