+++ ms2deepscore/data_generators.py
@@ -5,7 +5,7 @@ from typing import Iterator, List, NamedTuple, Optional
 import numpy as np
 import pandas as pd
 from tensorflow.keras.utils import Sequence  # pylint: disable=import-error
-from ms2deepscore.spectrum_pair_selection import SelectedCompoundPairs
+from ms2deepscore.train_new_model.spectrum_pair_selection import SelectedCompoundPairs
 from ms2deepscore.SpectrumBinner import SpectrumBinner
 from .typing import BinnedSpectrumType
